requirements.txt
Flask~=3.1.0
requests~=2.32.3
python-telegram-bot~=22.0
vk_api~=11.9.9
numpy~=2.2.5
wikipedia~=1.4.0
SQLAlchemy~=2.0.29
Flask-RESTful~=0.3.10
Flask-Login~=0.6.3
Werkzeug~=3.1.3
WTForms~=3.2.1